+++ dalesbred/src/main/java/fi/evident/dalesbred/Propagation.java
@@ -27,6 +27,9 @@ package fi.evident.dalesbred;
  */
 public enum Propagation {
 
+    /** Use the default propagation level that is configured */
+    DEFAULT,
+
     /** Join existing transaction if there is one, otherwise create a new one. */
     REQUIRED,
 
@@ -37,5 +40,11 @@ public enum Propagation {
     REQUIRES_NEW,
 
     /** Start a nested transaction if there is a current transaction, otherwise start a new normal transaction. */
-    NESTED,
+    NESTED,;
+
+    Propagation normalize(Propagation defaultValue) {
+        return (this != DEFAULT)         ? this
+             : (defaultValue != DEFAULT) ? defaultValue
+             : REQUIRED;
+    }
 }